Change MusicNote name, move interfaces in separate folder
