+++ app/src/main/java/com/app/music_app/MainActivity.kt
@@ -3,13 +3,17 @@ package com.app.music_app
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Button
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import com.app.music_app.instruments.VirtualPiano
 import com.app.music_app.view.piano_keyboard.PianoKeyboard
+import com.musiclib.Alteration
 import com.musiclib.notes.Note
 import com.musiclib.notes.NoteName
 import com.musiclib.notes.NoteRange
+import kotlinx.coroutines.GlobalScope
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -19,19 +23,18 @@ class MainActivity : ComponentActivity() {
         fun soundOf(note: Note) = vp.soundOf(this, note)
         setContent {
 
-            PianoKeyboard(
-                NoteRange(Note(NoteName.Do), Note(NoteName.Si)),
-                soundOf = ::soundOf
-            ).Draw()
+            Column {
+                val piano = PianoKeyboard(
+                    NoteRange(Note(NoteName.Si), Note(NoteName.Si, octave = 1)),
+                    soundOf = ::soundOf,
+                    context = LocalContext.current
+                )
+                piano.Draw()
+
+                Button(onClick = { piano.mark(Note(NoteName.Si)) }) {
+
+                }
+            }
         }
     }
-}
-
-
-@Preview
-@Composable
-fun PianoKeyboardPreview() {
-    PianoKeyboard(
-        NoteRange(Note(NoteName.Do), Note(NoteName.Si)),
-    ).Draw()
 }
