--- musiclib/src/main/java/com/musiclib/notes/MelodyNote.kt
@@ -1,6 +1,0 @@
-package com.musiclib.notes
-
-interface MelodyNote : MusicNote, MusicPause {
-    /** Динамика */
-    val volume: NoteVolume;
-}
