+++ musiclib/src/main/java/com/musiclib/notes/interfaces/MelodyNote.kt
@@ -1,0 +1,8 @@
+package com.musiclib.notes.interfaces
+
+import com.musiclib.notes.NoteVolume
+
+interface MelodyNote : BasicNote, MusicPause {
+    /** Динамика */
+    val volume: NoteVolume
+}
