+++ musiclib/src/main/java/com/musiclib/notes/interfaces/MusicPause.kt
@@ -1,4 +1,6 @@
-package com.musiclib.notes
+package com.musiclib.notes.interfaces
+
+import com.musiclib.notes.NoteDuration
 
 /** Представляет собой музыкальную паузу */
 interface MusicPause {
